Update docs, files name
